# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Jobs management service API in Go using Gin framework.",
        "title": "Gin Go Jobs Service",
        "contact": {
            "name": "David Afdal"
        },
        "version": "1.0"
    },
    "host": "103.157.26.181:7426",
    "basePath": "/api",
    "paths": {
        "/job": {
            "get": {
                "description": "Api Enpoint to Get All Data Jobs",
                "tags": [
                    "Jobs"
                ],
                "summary": "Get All Jobs.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Api Endpoint To Create New Job.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Create New Job",
                "parameters": [
                    {
                        "description": "Data Input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateJobInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    }
                }
            }
        },
        "/job/applier/{userId}": {
            "get": {
                "description": "Api endpoint to get applied job by userId.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Get Applied Jobs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "param enpoint",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    }
                }
            }
        },
        "/job/apply": {
            "post": {
                "description": "Api endpoint to apply job.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Applicants"
                ],
                "summary": "Apply job",
                "parameters": [
                    {
                        "description": "data input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ApplierRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    }
                }
            }
        },
        "/job/cancel": {
            "post": {
                "description": "Api Enpoint to cancel apply job.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Applicants"
                ],
                "summary": "Cancel Apply job",
                "parameters": [
                    {
                        "description": "data input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ApplierRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    }
                }
            }
        },
        "/job/{jobId}": {
            "get": {
                "description": "Api Enpoint to Get Single Job By Id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Get Single Job By Id.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Param Endpoint",
                        "name": "jobId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Api Endpoint for Delete job.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Delete job",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "param endpoint",
                        "name": "jobId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Api Endpoint To Update Job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Update Job",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Param Enpoint",
                        "name": "jobId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data Input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateJobInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    }
                }
            }
        },
        "/job/{userId}": {
            "get": {
                "description": "Api Endpoint To Get Shared Jobs.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Get Shared jobs by userId.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Param Endpoint",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.ApplierRequest": {
            "type": "object",
            "required": [
                "job_id",
                "user_id"
            ],
            "properties": {
                "job_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "request.CreateJobInput": {
            "type": "object",
            "required": [
                "category",
                "company",
                "description",
                "image",
                "location",
                "salary",
                "status",
                "title",
                "userId"
            ],
            "properties": {
                "category": {
                    "type": "string"
                },
                "company": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "salary": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "request.UpdateJobInput": {
            "type": "object",
            "required": [
                "category",
                "company",
                "description",
                "image",
                "location",
                "salary",
                "status",
                "title"
            ],
            "properties": {
                "category": {
                    "type": "string"
                },
                "company": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "salary": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "response.WebResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}